% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_associate_role_to_group}
\alias{greengrass_associate_role_to_group}
\title{Associates a role with a group}
\usage{
greengrass_associate_role_to_group(GroupId, RoleArn)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{RoleArn}{The ARN of the role you wish to associate with this group.}
}
\description{
Associates a role with a group. Your Greengrass core will use the role to access AWS cloud services. The role's permissions should allow Greengrass core Lambda functions to perform actions against the cloud.
}
\section{Request syntax}{
\preformatted{svc$associate_role_to_group(
  GroupId = "string",
  RoleArn = "string"
)
}
}

\keyword{internal}
