% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_create_subscription_definition_version}
\alias{greengrass_create_subscription_definition_version}
\title{Creates a version of a subscription definition which has already been
defined}
\usage{
greengrass_create_subscription_definition_version(AmznClientToken,
  SubscriptionDefinitionId, Subscriptions)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{SubscriptionDefinitionId}{[required] The ID of the subscription definition.}

\item{Subscriptions}{A list of subscriptions.}
}
\description{
Creates a version of a subscription definition which has already been defined.
}
\section{Request syntax}{
\preformatted{svc$create_subscription_definition_version(
  AmznClientToken = "string",
  SubscriptionDefinitionId = "string",
  Subscriptions = list(
    list(
      Id = "string",
      Source = "string",
      Subject = "string",
      Target = "string"
    )
  )
)
}
}

\keyword{internal}
