% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_disassociate_role_from_group}
\alias{greengrass_disassociate_role_from_group}
\title{Disassociates the role from a group}
\usage{
greengrass_disassociate_role_from_group(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Disassociates the role from a group.
}
\section{Request syntax}{
\preformatted{svc$disassociate_role_from_group(
  GroupId = "string"
)
}
}

\keyword{internal}
