% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_connectivity_info}
\alias{greengrass_get_connectivity_info}
\title{Retrieves the connectivity information for a core}
\usage{
greengrass_get_connectivity_info(ThingName)
}
\arguments{
\item{ThingName}{[required] The thing name.}
}
\description{
Retrieves the connectivity information for a core.
}
\section{Request syntax}{
\preformatted{svc$get_connectivity_info(
  ThingName = "string"
)
}
}

\keyword{internal}
