% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_device_definition}
\alias{greengrass_get_device_definition}
\title{Retrieves information about a device definition}
\usage{
greengrass_get_device_definition(DeviceDefinitionId)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}
}
\description{
Retrieves information about a device definition.
}
\section{Request syntax}{
\preformatted{svc$get_device_definition(
  DeviceDefinitionId = "string"
)
}
}

\keyword{internal}
