% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_function_definition}
\alias{greengrass_get_function_definition}
\title{Retrieves information about a Lambda function definition, including its
creation time and latest version}
\usage{
greengrass_get_function_definition(FunctionDefinitionId)
}
\arguments{
\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}
}
\description{
Retrieves information about a Lambda function definition, including its creation time and latest version.
}
\section{Request syntax}{
\preformatted{greengrass$get_function_definition(
  FunctionDefinitionId = "string"
)
}
}

\keyword{internal}
