% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_function_definition_version}
\alias{greengrass_get_function_definition_version}
\title{Retrieves information about a Lambda function definition version,
including which Lambda functions are included in the version and their
configurations}
\usage{
greengrass_get_function_definition_version(FunctionDefinitionId,
  FunctionDefinitionVersionId, NextToken)
}
\arguments{
\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}

\item{FunctionDefinitionVersionId}{[required] The ID of the function definition version.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations.
}
\section{Request syntax}{
\preformatted{svc$get_function_definition_version(
  FunctionDefinitionId = "string",
  FunctionDefinitionVersionId = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
