% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_group_certificate_authority}
\alias{greengrass_get_group_certificate_authority}
\title{Retreives the CA associated with a group}
\usage{
greengrass_get_group_certificate_authority(CertificateAuthorityId,
  GroupId)
}
\arguments{
\item{CertificateAuthorityId}{[required] The ID of the certificate authority.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Retreives the CA associated with a group. Returns the public key of the CA.
}
\section{Request syntax}{
\preformatted{svc$get_group_certificate_authority(
  CertificateAuthorityId = "string",
  GroupId = "string"
)
}
}

\keyword{internal}
