% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_get_subscription_definition}
\alias{greengrass_get_subscription_definition}
\title{Retrieves information about a subscription definition}
\usage{
greengrass_get_subscription_definition(SubscriptionDefinitionId)
}
\arguments{
\item{SubscriptionDefinitionId}{[required] The ID of the subscription definition.}
}
\description{
Retrieves information about a subscription definition.
}
\section{Request syntax}{
\preformatted{svc$get_subscription_definition(
  SubscriptionDefinitionId = "string"
)
}
}

\keyword{internal}
