% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_core_definition_versions}
\alias{greengrass_list_core_definition_versions}
\title{Lists the versions of a core definition}
\usage{
greengrass_list_core_definition_versions(CoreDefinitionId, MaxResults,
  NextToken)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a core definition.
}
\section{Request syntax}{
\preformatted{svc$list_core_definition_versions(
  CoreDefinitionId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

\keyword{internal}
