% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_list_group_certificate_authorities}
\alias{greengrass_list_group_certificate_authorities}
\title{Retrieves the current CAs for a group}
\usage{
greengrass_list_group_certificate_authorities(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Retrieves the current CAs for a group.
}
\section{Request syntax}{
\preformatted{svc$list_group_certificate_authorities(
  GroupId = "string"
)
}
}

\keyword{internal}
