% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greengrass_operations.R
\name{greengrass_update_logger_definition}
\alias{greengrass_update_logger_definition}
\title{Updates a logger definition}
\usage{
greengrass_update_logger_definition(LoggerDefinitionId, Name)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a logger definition.
}
\section{Request syntax}{
\preformatted{svc$update_logger_definition(
  LoggerDefinitionId = "string",
  Name = "string"
)
}
}

\keyword{internal}
