% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_claim_devices_by_claim_code}
\alias{iot1clickdevicesservice_claim_devices_by_claim_code}
\title{Adds device(s) to your account (i}
\usage{
iot1clickdevicesservice_claim_devices_by_claim_code(ClaimCode)
}
\arguments{
\item{ClaimCode}{[required] The claim code, starting with "C-", as provided by the device
manufacturer.}
}
\description{
Adds device(s) to your account (i.e., claim one or more devices) if and
only if you received a claim code with the device(s).
}
\section{Request syntax}{
\preformatted{iot1clickdevicesservice$claim_devices_by_claim_code(
  ClaimCode = "string"
)
}
}

\keyword{internal}
