% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_create_placement}
\alias{iot1clickprojects_create_placement}
\title{Creates an empty placement}
\usage{
iot1clickprojects_create_placement(placementName, projectName,
  attributes)
}
\arguments{
\item{placementName}{[required] The name of the placement to be created.}

\item{projectName}{[required] The name of the project in which to create the placement.}

\item{attributes}{Optional user-defined key/value pairs providing contextual data (such as
location or function) for the placement.}
}
\description{
Creates an empty placement.
}
\section{Request syntax}{
\preformatted{svc$create_placement(
  placementName = "string",
  projectName = "string",
  attributes = list(
    "string"
  )
)
}
}

\keyword{internal}
