% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_delete_placement}
\alias{iot1clickprojects_delete_placement}
\title{Deletes a placement}
\usage{
iot1clickprojects_delete_placement(placementName, projectName)
}
\arguments{
\item{placementName}{[required] The name of the empty placement to delete.}

\item{projectName}{[required] The project containing the empty placement to delete.}
}
\description{
Deletes a placement. To delete a placement, it must not have any devices
associated with it.
}
\details{
When you delete a placement, all associated data becomes irretrievable.
}
\section{Request syntax}{
\preformatted{svc$delete_placement(
  placementName = "string",
  projectName = "string"
)
}
}

\keyword{internal}
