% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_describe_placement}
\alias{iot1clickprojects_describe_placement}
\title{Describes a placement in a project}
\usage{
iot1clickprojects_describe_placement(placementName, projectName)
}
\arguments{
\item{placementName}{[required] The name of the placement within a project.}

\item{projectName}{[required] The project containing the placement to be described.}
}
\description{
Describes a placement in a project.
}
\section{Request syntax}{
\preformatted{svc$describe_placement(
  placementName = "string",
  projectName = "string"
)
}
}

\keyword{internal}
