% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_describe_project}
\alias{iot1clickprojects_describe_project}
\title{Returns an object describing a project}
\usage{
iot1clickprojects_describe_project(projectName)
}
\arguments{
\item{projectName}{[required] The name of the project to be described.}
}
\description{
Returns an object describing a project.
}
\section{Request syntax}{
\preformatted{svc$describe_project(
  projectName = "string"
)
}
}

\keyword{internal}
