% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_create_certificate_from_csr}
\alias{iot_create_certificate_from_csr}
\title{Creates an X}
\usage{
iot_create_certificate_from_csr(certificateSigningRequest, setAsActive)
}
\arguments{
\item{certificateSigningRequest}{[required] The certificate signing request (CSR).}

\item{setAsActive}{Specifies whether the certificate is active.}
}
\description{
Creates an X.509 certificate using the specified certificate signing
request.
}
\details{
\strong{Note:} The CSR must include a public key that is either an RSA key
with a length of at least 2048 bits or an ECC key from NIST P-256 or
NIST P-384 curves.

\strong{Note:} Reusing the same certificate signing request (CSR) results in
a distinct certificate.

You can create multiple certificates in a batch by creating a directory,
copying multiple .csr files into that directory, and then specifying
that directory on the command line. The following commands show how to
create a batch of certificates given a batch of CSRs.

Assuming a set of CSRs are located inside of the directory
my-csr-directory:

On Linux and OS X, the command is:

\$ ls my-csr-directory/ \| xargs -I {} aws iot
create-certificate-from-csr \--certificate-signing-request
file://my-csr-directory/{}

This command lists all of the CSRs in my-csr-directory and pipes each
CSR file name to the aws iot create-certificate-from-csr AWS CLI command
to create a certificate for the corresponding CSR.

The aws iot create-certificate-from-csr part of the command can also be
run in parallel to speed up the certificate creation process:

\$ ls my-csr-directory/ \| xargs -P 10 -I {} aws iot
create-certificate-from-csr \--certificate-signing-request
file://my-csr-directory/{}

On Windows PowerShell, the command to create certificates for all CSRs
in my-csr-directory is:

\> ls -Name my-csr-directory \| \%{aws iot create-certificate-from-csr
\--certificate-signing-request file://my-csr-directory/\$\_}

On a Windows command prompt, the command to create certificates for all
CSRs in my-csr-directory is:

\> forfiles /p my-csr-directory /c "cmd /c aws iot
create-certificate-from-csr \--certificate-signing-request
file://\@path"
}
\section{Request syntax}{
\preformatted{svc$create_certificate_from_csr(
  certificateSigningRequest = "string",
  setAsActive = TRUE|FALSE
)
}
}

\keyword{internal}
