% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_delete_thing}
\alias{iot_delete_thing}
\title{Deletes the specified thing}
\usage{
iot_delete_thing(thingName, expectedVersion)
}
\arguments{
\item{thingName}{[required] The name of the thing to delete.}

\item{expectedVersion}{The expected version of the thing record in the registry. If the version
of the record in the registry does not match the expected version
specified in the request, the \code{DeleteThing} request is rejected with a
\code{VersionConflictException}.}
}
\description{
Deletes the specified thing. Returns successfully with no error if the
deletion is successful or you specify a thing that doesn't exist.
}
\section{Request syntax}{
\preformatted{svc$delete_thing(
  thingName = "string",
  expectedVersion = 123
)
}
}

\keyword{internal}
