% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_authorizer}
\alias{iot_describe_authorizer}
\title{Describes an authorizer}
\usage{
iot_describe_authorizer(authorizerName)
}
\arguments{
\item{authorizerName}{[required] The name of the authorizer to describe.}
}
\description{
Describes an authorizer.
}
\section{Request syntax}{
\preformatted{iot$describe_authorizer(
  authorizerName = "string"
)
}
}

\keyword{internal}
