% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_billing_group}
\alias{iot_describe_billing_group}
\title{Returns information about a billing group}
\usage{
iot_describe_billing_group(billingGroupName)
}
\arguments{
\item{billingGroupName}{[required] The name of the billing group.}
}
\description{
Returns information about a billing group.
}
\section{Request syntax}{
\preformatted{svc$describe_billing_group(
  billingGroupName = "string"
)
}
}

\keyword{internal}
