% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_index}
\alias{iot_describe_index}
\title{Describes a search index}
\usage{
iot_describe_index(indexName)
}
\arguments{
\item{indexName}{[required] The index name.}
}
\description{
Describes a search index.
}
\section{Request syntax}{
\preformatted{iot$describe_index(
  indexName = "string"
)
}
}

\keyword{internal}
