% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_scheduled_audit}
\alias{iot_describe_scheduled_audit}
\title{Gets information about a scheduled audit}
\usage{
iot_describe_scheduled_audit(scheduledAuditName)
}
\arguments{
\item{scheduledAuditName}{[required] The name of the scheduled audit whose information you want to get.}
}
\description{
Gets information about a scheduled audit.
}
\section{Request syntax}{
\preformatted{iot$describe_scheduled_audit(
  scheduledAuditName = "string"
)
}
}

\keyword{internal}
