% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_security_profile}
\alias{iot_describe_security_profile}
\title{Gets information about a Device Defender security profile}
\usage{
iot_describe_security_profile(securityProfileName)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile whose information you want to get.}
}
\description{
Gets information about a Device Defender security profile.
}
\section{Request syntax}{
\preformatted{svc$describe_security_profile(
  securityProfileName = "string"
)
}
}

\keyword{internal}
