% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing_registration_task}
\alias{iot_describe_thing_registration_task}
\title{Describes a bulk thing provisioning task}
\usage{
iot_describe_thing_registration_task(taskId)
}
\arguments{
\item{taskId}{[required] The task ID.}
}
\description{
Describes a bulk thing provisioning task.
}
\section{Request syntax}{
\preformatted{svc$describe_thing_registration_task(
  taskId = "string"
)
}
}

\keyword{internal}
