% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_describe_thing_type}
\alias{iot_describe_thing_type}
\title{Gets information about the specified thing type}
\usage{
iot_describe_thing_type(thingTypeName)
}
\arguments{
\item{thingTypeName}{[required] The name of the thing type.}
}
\description{
Gets information about the specified thing type.
}
\section{Request syntax}{
\preformatted{iot$describe_thing_type(
  thingTypeName = "string"
)
}
}

\keyword{internal}
