% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_detach_policy}
\alias{iot_detach_policy}
\title{Detaches a policy from the specified target}
\usage{
iot_detach_policy(policyName, target)
}
\arguments{
\item{policyName}{[required] The policy to detach.}

\item{target}{[required] The target from which the policy will be detached.}
}
\description{
Detaches a policy from the specified target.
}
\section{Request syntax}{
\preformatted{svc$detach_policy(
  policyName = "string",
  target = "string"
)
}
}

\keyword{internal}
