% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_registration_code}
\alias{iot_get_registration_code}
\title{Gets a registration code used to register a CA certificate with AWS IoT}
\usage{
iot_get_registration_code()
}
\description{
Gets a registration code used to register a CA certificate with AWS IoT.
}
\section{Request syntax}{
\preformatted{iot$get_registration_code()
}
}

\keyword{internal}
