% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_statistics}
\alias{iot_get_statistics}
\title{Gets statistics about things that match the specified query}
\usage{
iot_get_statistics(indexName, queryString, aggregationField,
  queryVersion)
}
\arguments{
\item{indexName}{The name of the index to search. The default value is \code{AWS_Things}.}

\item{queryString}{[required] The query used to search. You can specify "*" for the query string to
get the count of all indexed things in your AWS account.}

\item{aggregationField}{The aggregation field name. Currently not supported.}

\item{queryVersion}{The version of the query used to search.}
}
\description{
Gets statistics about things that match the specified query.
}
\section{Request syntax}{
\preformatted{svc$get_statistics(
  indexName = "string",
  queryString = "string",
  aggregationField = "string",
  queryVersion = "string"
)
}
}

\keyword{internal}
