% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_get_topic_rule}
\alias{iot_get_topic_rule}
\title{Gets information about the rule}
\usage{
iot_get_topic_rule(ruleName)
}
\arguments{
\item{ruleName}{[required] The name of the rule.}
}
\description{
Gets information about the rule.
}
\section{Request syntax}{
\preformatted{svc$get_topic_rule(
  ruleName = "string"
)
}
}

\keyword{internal}
