% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_job_executions_for_thing}
\alias{iot_list_job_executions_for_thing}
\title{Lists the job executions for the specified thing}
\usage{
iot_list_job_executions_for_thing(thingName, status, maxResults,
  nextToken)
}
\arguments{
\item{thingName}{[required] The thing name.}

\item{status}{An optional filter that lets you search for jobs that have the specified
status.}

\item{maxResults}{The maximum number of results to be returned per request.}

\item{nextToken}{The token to retrieve the next set of results.}
}
\description{
Lists the job executions for the specified thing.
}
\section{Request syntax}{
\preformatted{iot$list_job_executions_for_thing(
  thingName = "string",
  status = "QUEUED"|"IN_PROGRESS"|"SUCCEEDED"|"FAILED"|"TIMED_OUT"|"REJECTED"|"REMOVED"|"CANCELED",
  maxResults = 123,
  nextToken = "string"
)
}
}

\keyword{internal}
