% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_jobs}
\alias{iot_list_jobs}
\title{Lists jobs}
\usage{
iot_list_jobs(status, targetSelection, maxResults, nextToken,
  thingGroupName, thingGroupId)
}
\arguments{
\item{status}{An optional filter that lets you search for jobs that have the specified
status.}

\item{targetSelection}{Specifies whether the job will continue to run (CONTINUOUS), or will be
complete after all those things specified as targets have completed the
job (SNAPSHOT). If continuous, the job may also be run on a thing when a
change is detected in a target. For example, a job will run on a thing
when the thing is added to a target group, even after the job was
completed by all things originally in the group.}

\item{maxResults}{The maximum number of results to return per request.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{thingGroupName}{A filter that limits the returned jobs to those for the specified group.}

\item{thingGroupId}{A filter that limits the returned jobs to those for the specified group.}
}
\description{
Lists jobs.
}
\section{Request syntax}{
\preformatted{svc$list_jobs(
  status = "IN_PROGRESS"|"CANCELED"|"COMPLETED"|"DELETION_IN_PROGRESS",
  targetSelection = "CONTINUOUS"|"SNAPSHOT",
  maxResults = 123,
  nextToken = "string",
  thingGroupName = "string",
  thingGroupId = "string"
)
}
}

\keyword{internal}
