% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_outgoing_certificates}
\alias{iot_list_outgoing_certificates}
\title{Lists certificates that are being transferred but not yet accepted}
\usage{
iot_list_outgoing_certificates(pageSize, marker, ascendingOrder)
}
\arguments{
\item{pageSize}{The result page size.}

\item{marker}{The marker for the next set of results.}

\item{ascendingOrder}{Specifies the order for results. If True, the results are returned in
ascending order, based on the creation date.}
}
\description{
Lists certificates that are being transferred but not yet accepted.
}
\section{Request syntax}{
\preformatted{iot$list_outgoing_certificates(
  pageSize = 123,
  marker = "string",
  ascendingOrder = TRUE|FALSE
)
}
}

\keyword{internal}
