% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_policies}
\alias{iot_list_policies}
\title{Lists your policies}
\usage{
iot_list_policies(marker, pageSize, ascendingOrder)
}
\arguments{
\item{marker}{The marker for the next set of results.}

\item{pageSize}{The result page size.}

\item{ascendingOrder}{Specifies the order for results. If true, the results are returned in
ascending creation order.}
}
\description{
Lists your policies.
}
\section{Request syntax}{
\preformatted{svc$list_policies(
  marker = "string",
  pageSize = 123,
  ascendingOrder = TRUE|FALSE
)
}
}

\keyword{internal}
