% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_thing_groups_for_thing}
\alias{iot_list_thing_groups_for_thing}
\title{List the thing groups to which the specified thing belongs}
\usage{
iot_list_thing_groups_for_thing(thingName, nextToken, maxResults)
}
\arguments{
\item{thingName}{[required] The thing name.}

\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
List the thing groups to which the specified thing belongs.
}
\section{Request syntax}{
\preformatted{iot$list_thing_groups_for_thing(
  thingName = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
