% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_thing_principals}
\alias{iot_list_thing_principals}
\title{Lists the principals associated with the specified thing}
\usage{
iot_list_thing_principals(thingName)
}
\arguments{
\item{thingName}{[required] The name of the thing.}
}
\description{
Lists the principals associated with the specified thing. A principal
can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito
identities or federated identities.
}
\section{Request syntax}{
\preformatted{iot$list_thing_principals(
  thingName = "string"
)
}
}

\keyword{internal}
