% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_thing_registration_tasks}
\alias{iot_list_thing_registration_tasks}
\title{List bulk thing provisioning tasks}
\usage{
iot_list_thing_registration_tasks(nextToken, maxResults, status)
}
\arguments{
\item{nextToken}{The token to retrieve the next set of results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{status}{The status of the bulk thing provisioning task.}
}
\description{
List bulk thing provisioning tasks.
}
\section{Request syntax}{
\preformatted{iot$list_thing_registration_tasks(
  nextToken = "string",
  maxResults = 123,
  status = "InProgress"|"Completed"|"Failed"|"Cancelled"|"Cancelling"
)
}
}

\keyword{internal}
