% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_topic_rules}
\alias{iot_list_topic_rules}
\title{Lists the rules for the specific topic}
\usage{
iot_list_topic_rules(topic, maxResults, nextToken, ruleDisabled)
}
\arguments{
\item{topic}{The topic.}

\item{maxResults}{The maximum number of results to return.}

\item{nextToken}{A token used to retrieve the next value.}

\item{ruleDisabled}{Specifies whether the rule is disabled.}
}
\description{
Lists the rules for the specific topic.
}
\section{Request syntax}{
\preformatted{svc$list_topic_rules(
  topic = "string",
  maxResults = 123,
  nextToken = "string",
  ruleDisabled = TRUE|FALSE
)
}
}

\keyword{internal}
