% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_list_v2_logging_levels}
\alias{iot_list_v2_logging_levels}
\title{Lists logging levels}
\usage{
iot_list_v2_logging_levels(targetType, nextToken, maxResults)
}
\arguments{
\item{targetType}{The type of resource for which you are configuring logging. Must be
\code{THING_Group}.}

\item{nextToken}{The token used to get the next set of results, or \strong{null} if there are
no additional results.}

\item{maxResults}{The maximum number of results to return at one time.}
}
\description{
Lists logging levels.
}
\section{Request syntax}{
\preformatted{iot$list_v2_logging_levels(
  targetType = "DEFAULT"|"THING_GROUP",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
