% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_remove_thing_from_billing_group}
\alias{iot_remove_thing_from_billing_group}
\title{Removes the given thing from the billing group}
\usage{
iot_remove_thing_from_billing_group(billingGroupName, billingGroupArn,
  thingName, thingArn)
}
\arguments{
\item{billingGroupName}{The name of the billing group.}

\item{billingGroupArn}{The ARN of the billing group.}

\item{thingName}{The name of the thing to be removed from the billing group.}

\item{thingArn}{The ARN of the thing to be removed from the billing group.}
}
\description{
Removes the given thing from the billing group.
}
\section{Request syntax}{
\preformatted{svc$remove_thing_from_billing_group(
  billingGroupName = "string",
  billingGroupArn = "string",
  thingName = "string",
  thingArn = "string"
)
}
}

\keyword{internal}
