% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_search_index}
\alias{iot_search_index}
\title{The query search index}
\usage{
iot_search_index(indexName, queryString, nextToken, maxResults,
  queryVersion)
}
\arguments{
\item{indexName}{The search index name.}

\item{queryString}{[required] The search query string.}

\item{nextToken}{The token used to get the next set of results, or null if there are no
additional results.}

\item{maxResults}{The maximum number of results to return at one time.}

\item{queryVersion}{The query version.}
}
\description{
The query search index.
}
\section{Request syntax}{
\preformatted{svc$search_index(
  indexName = "string",
  queryString = "string",
  nextToken = "string",
  maxResults = 123,
  queryVersion = "string"
)
}
}

\keyword{internal}
