% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_dynamic_thing_group}
\alias{iot_update_dynamic_thing_group}
\title{Updates a dynamic thing group}
\usage{
iot_update_dynamic_thing_group(thingGroupName, thingGroupProperties,
  expectedVersion, indexName, queryString, queryVersion)
}
\arguments{
\item{thingGroupName}{[required] The name of the dynamic thing group to update.}

\item{thingGroupProperties}{[required] The dynamic thing group properties to update.}

\item{expectedVersion}{The expected version of the dynamic thing group to update.}

\item{indexName}{The dynamic thing group index to update.

Currently one index is supported: 'AWS\_Things'.}

\item{queryString}{The dynamic thing group search query string to update.}

\item{queryVersion}{The dynamic thing group query version to update.

Currently one query version is supported: "2017-09-30". If not
specified, the query version defaults to this value.}
}
\description{
Updates a dynamic thing group.
}
\section{Request syntax}{
\preformatted{svc$update_dynamic_thing_group(
  thingGroupName = "string",
  thingGroupProperties = list(
    thingGroupDescription = "string",
    attributePayload = list(
      attributes = list(
        "string"
      ),
      merge = TRUE|FALSE
    )
  ),
  expectedVersion = 123,
  indexName = "string",
  queryString = "string",
  queryVersion = "string"
)
}
}

\keyword{internal}
