% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot_operations.R
\name{iot_update_security_profile}
\alias{iot_update_security_profile}
\title{Updates a Device Defender security profile}
\usage{
iot_update_security_profile(securityProfileName,
  securityProfileDescription, behaviors, alertTargets,
  additionalMetricsToRetain, deleteBehaviors, deleteAlertTargets,
  deleteAdditionalMetricsToRetain, expectedVersion)
}
\arguments{
\item{securityProfileName}{[required] The name of the security profile you want to update.}

\item{securityProfileDescription}{A description of the security profile.}

\item{behaviors}{Specifies the behaviors that, when violated by a device (thing), cause
an alert.}

\item{alertTargets}{Where the alerts are sent. (Alerts are always sent to the console.)}

\item{additionalMetricsToRetain}{A list of metrics whose data is retained (stored). By default, data is
retained for any metric used in the profile's \code{behaviors} but it is
also retained for any metric specified here.}

\item{deleteBehaviors}{If true, delete all \code{behaviors} defined for this security profile. If
any \code{behaviors} are defined in the current invocation an exception
occurs.}

\item{deleteAlertTargets}{If true, delete all \code{alertTargets} defined for this security profile. If
any \code{alertTargets} are defined in the current invocation an exception
occurs.}

\item{deleteAdditionalMetricsToRetain}{If true, delete all \code{additionalMetricsToRetain} defined for this
security profile. If any \code{additionalMetricsToRetain} are defined in the
current invocation an exception occurs.}

\item{expectedVersion}{The expected version of the security profile. A new version is generated
whenever the security profile is updated. If you specify a value that is
different than the actual version, a \code{VersionConflictException} is
thrown.}
}
\description{
Updates a Device Defender security profile.
}
\section{Request syntax}{
\preformatted{svc$update_security_profile(
  securityProfileName = "string",
  securityProfileDescription = "string",
  behaviors = list(
    list(
      name = "string",
      metric = "string",
      criteria = list(
        comparisonOperator = "less-than"|"less-than-equals"|"greater-than"|"greater-than-equals"|"in-cidr-set"|"not-in-cidr-set"|"in-port-set"|"not-in-port-set",
        value = list(
          count = 123,
          cidrs = list(
            "string"
          ),
          ports = list(
            123
          )
        ),
        durationSeconds = 123,
        consecutiveDatapointsToAlarm = 123,
        consecutiveDatapointsToClear = 123,
        statisticalThreshold = list(
          statistic = "string"
        )
      )
    )
  ),
  alertTargets = list(
    list(
      alertTargetArn = "string",
      roleArn = "string"
    )
  ),
  additionalMetricsToRetain = list(
    "string"
  ),
  deleteBehaviors = TRUE|FALSE,
  deleteAlertTargets = TRUE|FALSE,
  deleteAdditionalMetricsToRetain = TRUE|FALSE,
  expectedVersion = 123
)
}
}

\keyword{internal}
