% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_delete_channel}
\alias{iotanalytics_delete_channel}
\title{Deletes the specified channel}
\usage{
iotanalytics_delete_channel(channelName)
}
\arguments{
\item{channelName}{[required] The name of the channel to delete.}
}
\description{
Deletes the specified channel.
}
\section{Request syntax}{
\preformatted{svc$delete_channel(
  channelName = "string"
)
}
}

\keyword{internal}
