% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_describe_logging_options}
\alias{iotanalytics_describe_logging_options}
\title{Retrieves the current settings of the AWS IoT Analytics logging options}
\usage{
iotanalytics_describe_logging_options()
}
\description{
Retrieves the current settings of the AWS IoT Analytics logging options.
}
\section{Request syntax}{
\preformatted{iotanalytics$describe_logging_options()
}
}

\keyword{internal}
