% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_describe_pipeline}
\alias{iotanalytics_describe_pipeline}
\title{Retrieves information about a pipeline}
\usage{
iotanalytics_describe_pipeline(pipelineName)
}
\arguments{
\item{pipelineName}{[required] The name of the pipeline whose information is retrieved.}
}
\description{
Retrieves information about a pipeline.
}
\section{Request syntax}{
\preformatted{svc$describe_pipeline(
  pipelineName = "string"
)
}
}

\keyword{internal}
