% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_list_pipelines}
\alias{iotanalytics_list_pipelines}
\title{Retrieves a list of pipelines}
\usage{
iotanalytics_list_pipelines(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in this request.

The default value is 100.}
}
\description{
Retrieves a list of pipelines.
}
\section{Request syntax}{
\preformatted{iotanalytics$list_pipelines(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
