% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotanalytics_operations.R
\name{iotanalytics_untag_resource}
\alias{iotanalytics_untag_resource}
\title{Removes the given tags (metadata) from the resource}
\usage{
iotanalytics_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags you want to remove.}

\item{tagKeys}{[required] The keys of those tags which you want to remove.}
}
\description{
Removes the given tags (metadata) from the resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
