% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdataplane_operations.R
\name{iotdataplane_update_thing_shadow}
\alias{iotdataplane_update_thing_shadow}
\title{Updates the thing shadow for the specified thing}
\usage{
iotdataplane_update_thing_shadow(thingName, payload)
}
\arguments{
\item{thingName}{[required] The name of the thing.}

\item{payload}{[required] The state information, in JSON format.}
}
\description{
Updates the thing shadow for the specified thing.
}
\details{
For more information, see
\href{http://docs.aws.amazon.com/iot/latest/developerguide/API_UpdateThingShadow.html}{UpdateThingShadow}
in the \emph{AWS IoT Developer Guide}.
}
\section{Request syntax}{
\preformatted{iotdataplane$update_thing_shadow(
  thingName = "string",
  payload = raw
)
}
}

\keyword{internal}
