% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotjobsdataplane_operations.R
\name{iotjobsdataplane_get_pending_job_executions}
\alias{iotjobsdataplane_get_pending_job_executions}
\title{Gets the list of all jobs for a thing that are not in a terminal status}
\usage{
iotjobsdataplane_get_pending_job_executions(thingName)
}
\arguments{
\item{thingName}{[required] The name of the thing that is executing the job.}
}
\description{
Gets the list of all jobs for a thing that are not in a terminal status.
}
\section{Request syntax}{
\preformatted{iotjobsdataplane$get_pending_job_executions(
  thingName = "string"
)
}
}

\keyword{internal}
