% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signer_operations.R
\name{signer_get_signing_profile}
\alias{signer_get_signing_profile}
\title{Returns information on a specific signing profile}
\usage{
signer_get_signing_profile(profileName)
}
\arguments{
\item{profileName}{[required] The name of the target signing profile.}
}
\description{
Returns information on a specific signing profile.
}
\section{Request syntax}{
\preformatted{signer$get_signing_profile(
  profileName = "string"
)
}
}

\keyword{internal}
